# docs/openapi.yaml
openapi: 3.1.0
info:
  title: kb-partners API
  version: 0.1.0
  description: >-
    Модуль обработки партнёрских заявок reHome (ТЗ v1.1). Contract-first:
    эта спецификация — источник истины для контрактных тестов (AT-002).
    На M0 (каркас) определены только инфраструктурные эндпоинты; доменные
    пути (/requests, /channels, /inbound, /webhooks) добавляются по эпикам
    M1–M7 под префиксом /api/v1/partners.
  contact:
    name: ООО РЕХОМ
servers:
  - url: https://api.rehome.one
    description: Production
  - url: https://staging-api.rehome.one
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/partners/requests:
    get:
      tags: [Requests]
      summary: Список заявок (scope-фильтр, курсорная пагинация)
      operationId: listRequests
      description: >-
        Видимые субъекту заявки (контур access_level + владение). Курсорная
        пагинация; фильтры status / category / partner_id / период (§11.1).
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RequestStatus"
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Category"
        - name: partner_id
          in: query
          required: false
          schema:
            type: string
        - name: created_from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: created_to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Страница заявок.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"
        "401":
          description: Не аутентифицирован.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Requests]
      summary: Приём заявки (ЛК-форма / m2m)
      operationId: createRequest
      description: >-
        FR-1.1/FR-1.4. `channel_in` выводится бэкендом из типа субъекта (не из тела).
        Идемпотентность — по заголовку `Idempotency-Key`.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Ключ дедупликации приёма; повтор возвращает существующую заявку (200).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCreate"
      responses:
        "201":
          description: Заявка создана.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRead"
        "200":
          description: Идемпотентный повтор — существующая заявка.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRead"
        "401":
          description: Не аутентифицирован.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Ошибка валидации тела.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/from-chat:
    post:
      tags: [Requests]
      summary: Приём заявки из AI-чата (kb-search, SERVICE-only)
      operationId: createRequestFromChat
      description: FR-1.2. Только сервис-принципал (m2m). Идемпотентность по `chat_session_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FromChatCreate"
      responses:
        "201":
          description: Заявка создана.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRead"
        "200":
          description: Идемпотентный повтор — существующая заявка.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRead"
        "403":
          description: Требуется сервис-принципал.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/from-ticket:
    post:
      tags: [Requests]
      summary: Эскалация заявки из тикета (kb-support, SERVICE-only)
      operationId: createRequestFromTicket
      description: FR-1.3. Только сервис-принципал (m2m). Идемпотентность по `ticket_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FromTicketCreate"
      responses:
        "201":
          description: Заявка создана.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRead"
        "200":
          description: Идемпотентный повтор — существующая заявка.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRead"
        "403":
          description: Требуется сервис-принципал.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Requests]
      summary: Карточка заявки + allowed_transitions
      operationId: getRequest
      description: >-
        Карточка с `allowed_transitions` (§7). `raw_input` маскируется по scope.
        Недоступный ресурс → 404 (анти-enumeration).
      responses:
        "200":
          description: Карточка заявки.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/transition:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Переход статуса (валидация по FSM)
      operationId: transitionRequest
      description: Переход FSM (§7). Запрещённый → 409; недоступная → 404; нет прав → 403.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Новое состояние карточки.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Действие недоступно субъекту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Запрещённый переход (FSM).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/cancel:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Отмена заявки (с причиной)
      operationId: cancelRequest
      description: Отмена из нетерминального статуса (§7). Партнёру запрещена (403).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRequest"
      responses:
        "200":
          description: Заявка отменена.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Действие недоступно субъекту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Заявка уже в терминальном статусе.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/classify:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: (Ре)классификация категории (operator/agent)
      operationId: classifyRequest
      description: >-
        E2: rules + LLM по `raw_input_masked` (FR-2.1–2.3); ниже порога уверенности
        или неоднозначно → NEEDS_REVIEW (FR-2.4). Трассировка в `classification`.
      responses:
        "200":
          description: Карточка с обновлённой категорией/классификацией.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Доступно только оператору/агенту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Классификация недопустима в текущем статусе.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/assign:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Подбор / назначение партнёра (operator/agent)
      operationId: assignRequest
      description: >-
        E3 (FR-3.1–3.4). partner_id → ручное назначение; пусто → авто-подбор по
        реестру kb-platform (ранжирование + fallback_chain + match_trace).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRequest"
      responses:
        "200":
          description: Назначено; карточка с partner_id/каналом/трассировкой.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Действие недоступно субъекту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Назначение недопустимо в текущем статусе / заявка не классифицирована.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Нет пригодных партнёров для авто-подбора.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/accept:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Приёмка пользователем
      operationId: acceptRequest
      description: FR-7.1. DONE→ACCEPTED_BY_USER. Партнёру недоступно (403); не DONE → 409.
      responses:
        "200":
          description: Заявка принята пользователем.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Действие недоступно субъекту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Приёмка недопустима в текущем статусе.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/dispute:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Открыть спор (с причиной)
      operationId: disputeRequest
      description: >-
        FR-7.2. DONE|ACCEPTED_BY_USER→DISPUTE; при сконфигурированном контуре
        порождает претензию COMPENSATION в kb-support (claim_ref).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeRequest"
      responses:
        "200":
          description: Спор открыт.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Действие недоступно субъекту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Спор недопустим в текущем статусе.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/settlement:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Подтверждение расчёта контуром (SERVICE-only)
      operationId: confirmSettlement
      description: >-
        FR-7.3. Платёжный контур rehome.one подтверждает расчёт →
        ACCEPTED_BY_USER→PAID. Деньги модуль не считает — хранит ссылки.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementConfirm"
      responses:
        "200":
          description: Заявка оплачена (PAID).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Требуется сервис-принципал.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Расчёт недопустим в текущем статусе.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/dispatch:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Диспетчеризация партнёру (operator/agent)
      operationId: dispatchRequest
      description: >-
        E4 (FR-4.1–4.6). Доставка по каналу с наивысшим priority среди активных;
        fallback по партнёрам цепочки; исчерпание → FAILED_DISPATCH. Каждая попытка —
        DispatchAttempt.
      responses:
        "200":
          description: Результат диспетчеризации (DISPATCHED или FAILED_DISPATCH).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Действие недоступно субъекту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Диспетчеризация недопустима в текущем статусе.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/partner-response:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Requests]
      summary: Ответ партнёра (портал LIGHT)
      operationId: partnerResponse
      description: >-
        E10/FR-10.2. Партнёр accepted/rejected/in_progress/done по своей заявке.
        Только партнёр (403); чужая → 404; запрещённый переход → 409; неизвестный → 422.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerResponse"
      responses:
        "200":
          description: Новое состояние карточки.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "403":
          description: Доступно только партнёру-исполнителю.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Запрещённый переход (FSM).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Неизвестный статус партнёра.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/requests/{request_id}/requester-context:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Requests]
      summary: Контекст заявителя из rehome.one (operator/agent)
      operationId: getRequesterContext
      description: >-
        E9/FR-9.1. User/Premises/Booking заявителя (контекст для оператора/агента).
        Только привилегированные роли (403); недоступная заявка → 404.
      responses:
        "200":
          description: Контекст заявителя (поля опциональны / контур недоступен → null).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequesterContextResponse"
        "403":
          description: Доступно только оператору/агенту.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заявка не найдена или недоступна.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak RS256 Bearer JWT (aud содержит `kb-partners`).
  schemas:
...
    Error:
      description: RFC 7807 problem+json (без ПДн в detail).
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties:
              type: string

    # --- Web-push подписки (E8, FR-10.1) ---
...
    Category:
      type: string
      enum: [CLEANING, MOVING, REPAIR, KEY_DELIVERY, OTHER]
    RequestStatus:
      type: string
      enum:
        [
          NEW,
          CLASSIFYING,
          CLASSIFIED,
          NEEDS_REVIEW,
          MATCHING,
          ASSIGNED,
          DISPATCHED,
          FAILED_DISPATCH,
          ACCEPTED,
          IN_PROGRESS,
          DONE,
          ACCEPTED_BY_USER,
          DISPUTE,
          PAID,
          CANCELLED,
          REJECTED,
        ]
    RequestCreate:
      description: >-
        Тело приёма из ЛК-формы / m2m. `channel_in` НЕ принимается (выводится
        бэкендом); `raw_input` маскируется до логов/LLM (FR-1.6).
      type: object
      additionalProperties: false
      required: [raw_input]
      properties:
        raw_input:
          type: string
          minLength: 1
          maxLength: 20000
        requester_id:
          type: string
          maxLength: 255
        booking_id:
          type: string
          maxLength: 255
        premises_id:
          type: string
          maxLength: 255
        source_ref:
          type: object
          additionalProperties: true
...
    RequestRead:
      description: Представление заявки наружу. Без raw_input (ПДн, FR-1.6/NFR-5).
      type: object
      additionalProperties: false
      required: [id, number, requester_id, channel_in, category, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        requester_id:
          type: string
        channel_in:
          $ref: "#/components/schemas/ChannelIn"
        category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
        status:
          $ref: "#/components/schemas/RequestStatus"
        created_at:
          type: string
          format: date-time
...
    TransitionRequest:
      description: Переход FSM; target валидируется по §7.
      type: object
      additionalProperties: false
      required: [target]
      properties:
        target:
          $ref: "#/components/schemas/RequestStatus"
...
    RequestListResponse:
      description: Страница списка заявок + курсор следующей страницы (или null).
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RequestRead"
        next_cursor:
          oneOf:
            - type: string
            - type: "null"

    # --- Каналы (§6.4, §11.2) ---
...